chore: Moved wallet icons to svg folder
