+++ src/components/Svg/WalletsIcons/Bitkeep.tsx
@@ -1,10 +1,9 @@
 import React from "react";
-import LegacySvg from "../../../components/Svg/LegacySvg";
-import { LegacySvgProps } from "../../../components/Svg/types";
+import { SvgProps } from "../types";
 
-const Icon: React.FC<LegacySvgProps> = (props) => {
+const Icon: React.FC<SvgProps> = (props) => {
   return (
-    <LegacySvg viewBox="0 0 128 128" {...props}>
+    <svg viewBox="0 0 128 128" {...props}>
       <rect width="128" height="128" rx="64" fill="#7524F9" />
       <path
         fillRule="evenodd"
@@ -12,7 +11,7 @@ const Icon: React.FC<LegacySvgProps> = (props) => {
         d="M102 45.6027V49.8204C102.001 50.2283 101.893 50.6292 101.687 50.9827C101.481 51.3361 101.185 51.6296 100.828 51.8335L87.0908 59.6801L99.3633 66.671C100.165 67.1295 100.831 67.7892 101.294 68.5836C101.757 69.378 102.001 70.2791 102 71.1962V82.4424C102.001 83.36 101.758 84.2617 101.295 85.0567C100.832 85.8516 100.166 86.5117 99.3633 86.9705L67.0869 105.3C66.2853 105.759 65.3759 106 64.4502 106C63.5245 106 62.6151 105.759 61.8135 105.3L51.252 99.2611C51.0737 99.1593 50.9256 99.0128 50.8227 98.8363C50.7198 98.6599 50.6656 98.4597 50.6656 98.256C50.6656 98.0522 50.7198 97.852 50.8227 97.6756C50.9256 97.4991 51.0737 97.3527 51.252 97.2508L86.1152 77.3835C86.204 77.3325 86.2777 77.2593 86.329 77.1712C86.3802 77.0832 86.4072 76.9833 86.4072 76.8817C86.4072 76.78 86.3802 76.6802 86.329 76.5921C86.2777 76.504 86.204 76.4308 86.1152 76.3798L73.1719 68.9771C72.8156 68.7734 72.4114 68.6662 72 68.6662C71.5886 68.6662 71.1844 68.7734 70.8281 68.9771L35.417 89.1722C35.1498 89.325 34.8467 89.4054 34.5381 89.4054C34.2295 89.4054 33.9264 89.325 33.6592 89.1722L29.6484 86.9009C28.8424 86.4428 28.1729 85.782 27.7078 84.9853C27.2427 84.1886 26.9985 83.2843 27 82.3641V77.7664C26.9998 77.4607 27.0809 77.1603 27.2352 76.8955C27.3895 76.6307 27.6115 76.4109 27.8789 76.258L78.6504 47.3693C78.7392 47.3183 78.8129 47.2451 78.8641 47.157C78.9154 47.0689 78.9423 46.9691 78.9423 46.8674C78.9423 46.7658 78.9154 46.6659 78.8641 46.5779C78.8129 46.4898 78.7392 46.4166 78.6504 46.3656L65.6865 38.9367C65.3302 38.7331 64.9261 38.6258 64.5146 38.6258C64.1032 38.6258 63.6991 38.7331 63.3428 38.9367L28.7578 58.63C28.5798 58.7318 28.3778 58.7854 28.1722 58.7855C27.9666 58.7855 27.7646 58.732 27.5865 58.6303C27.4084 58.5286 27.2604 58.3823 27.1575 58.2061C27.0545 58.0299 27.0002 57.83 27 57.6264V45.5476C26.9989 44.63 27.242 43.7283 27.7049 42.9334C28.1678 42.1384 28.8341 41.4783 29.6367 41.0195L61.9072 22.6954C62.7071 22.2398 63.6138 22 64.5366 22C65.4594 22 66.3661 22.2398 67.166 22.6954L99.3633 41.0833C100.165 41.5414 100.83 42.2001 101.293 42.9934C101.756 43.7867 102 44.6866 102 45.6027Z"
         fill="white"
       />
-    </LegacySvg>
+    </svg>
   );
 };
