+++ src/components/Svg/WalletsIcons/MadWallet.tsx
@@ -1,10 +1,9 @@
 import React from "react";
-import LegacySvg from "../../../components/Svg/LegacySvg";
-import { LegacySvgProps } from "../../../components/Svg/types";
+import { SvgProps } from "../types";
 
-const Icon: React.FC<LegacySvgProps> = (props) => {
+const Icon: React.FC<SvgProps> = (props) => {
   return (
-    <LegacySvg viewBox="0 0 128 128" {...props}>
+    <svg viewBox="0 0 128 128" {...props}>
       <rect width="128" height="128" rx="64" fill="#0E7696" />
       <path
         d="M100.701 60.3101C102.279 59.4998 103.858 58.7106 105.426 57.8687C106.383 57.3531 106.478 57.774 106.478 58.6474C106.478 64.13 106.552 69.6231 106.478 75.1057C106.528 77.1473 106.021 79.1639 105.013 80.9397C104.004 82.7155 102.532 84.1837 100.753 85.187C92.64 90.1013 84.3582 94.6579 76.0764 99.3196C75.8238 99.4212 75.5634 99.5021 75.2977 99.5617C75.2977 97.7306 75.2977 96.0259 75.2977 94.3001C75.2977 93.2477 75.0031 93.2477 74.1717 93.6687C71.9418 95.016 69.6048 96.1775 67.1843 97.1413C65.8519 97.6534 64.4186 97.8486 62.9978 97.7117C61.5769 97.5747 60.2074 97.1092 58.9973 96.3521C57.1557 95.2472 55.272 94.2475 53.1674 93.0373V99.246C52.3571 99.467 51.9678 98.9724 51.5153 98.7198C43.9912 94.4474 36.4881 90.1434 28.964 85.892C24.418 83.3138 21.9976 79.5781 22.0608 74.3165C22.0608 68.7287 22.0608 63.1408 22.0608 57.2373L27.8591 60.2575C27.9834 60.4131 28.0828 60.587 28.1537 60.7731C28.2375 61.0603 28.28 61.3579 28.28 61.6571C28.28 66.1084 28.28 70.5492 28.28 74.99C28.2426 76.0735 28.5097 77.1458 29.0509 78.0851C29.5922 79.0245 30.386 79.7933 31.3423 80.3042C34.6255 82.2615 37.9403 84.1346 41.2867 85.9867C42.7178 86.9128 44.2542 87.6915 45.6959 88.6175C46.5378 89.1542 47.0113 88.9753 47.0218 87.8914C47.0324 86.8075 46.9692 85.7868 47.0218 84.7345C47.1376 83.3244 47.5585 83.0823 48.8213 83.7348C51.3153 85.0291 53.7041 86.5339 56.1876 87.944C57.8923 89.049 59.6707 90.0487 61.4492 90.9853C63.4907 92.0376 65.4796 91.5851 67.4053 90.5643C71.3515 88.7859 74.9084 86.355 78.7178 84.2504C79.4228 83.8716 80.1279 83.0192 80.9592 83.4717C81.7905 83.9242 81.4538 84.9555 81.4959 85.7657C81.538 86.576 81.4959 87.1969 81.4959 87.8704C81.4959 88.9227 81.98 88.9227 82.7271 88.4386C87.6204 85.7447 92.3664 82.7877 97.1965 79.9674C98.1916 79.4 99.0072 78.5645 99.5505 77.5561C100.094 76.5477 100.343 75.4069 100.269 74.2639C100.269 70.3387 100.269 66.403 100.269 62.4673C100.185 61.7209 100.336 60.9668 100.701 60.3101Z"
@@ -24,7 +23,7 @@ const Icon: React.FC<LegacySvgProps> = (props) => {
         fill="#261A41"
         fillOpacity="0.53"
       />
-    </LegacySvg>
+    </svg>
   );
 };
