+++ src/components/Svg/WalletsIcons/TorusWallet.tsx
@@ -1,10 +1,9 @@
 import React from "react";
-import LegacySvg from "../../../components/Svg/LegacySvg";
-import { LegacySvgProps } from "../../../components/Svg/types";
+import { SvgProps } from "../types";
 
-const Icon: React.FC<LegacySvgProps> = (props) => {
+const Icon: React.FC<SvgProps> = (props) => {
   return (
-    <LegacySvg viewBox="0 0 32 32" {...props}>
+    <svg viewBox="0 0 32 32" {...props}>
       <path d="M16 32c8.837 0 16-7.163 16-16S24.837 0 16 0 0 7.163 0 16s7.163 16 16 16Z" fill="#0364FF" />
       <path
         fillRule="evenodd"
@@ -13,7 +12,7 @@ const Icon: React.FC<LegacySvgProps> = (props) => {
         fill="#fff"
       />
       <path d="M20.838 13.876a2.191 2.191 0 1 0 0-4.382 2.191 2.191 0 0 0 0 4.382Z" fill="#fff" />
-    </LegacySvg>
+    </svg>
   );
 };
