+++ src/components/Svg/WalletsIcons/TrustWallet.tsx
@@ -1,10 +1,9 @@
 import React from "react";
-import LegacySvg from "../../../components/Svg/LegacySvg";
-import { LegacySvgProps } from "../../../components/Svg/types";
+import { SvgProps } from "../types";
 
-const Icon: React.FC<LegacySvgProps> = (props) => {
+const Icon: React.FC<SvgProps> = (props) => {
   return (
-    <LegacySvg viewBox="0 0 96 96" {...props}>
+    <svg viewBox="0 0 96 96" {...props}>
       <g clipPath="url(#clip0)">
         <path
           d="M48.0048 96.0097C74.5172 96.0097 96.0097 74.5172 96.0097 48.0048C96.0097 21.4925 74.5172 0 48.0048 0C21.4925 0 0 21.4925 0 48.0048C0 74.5172 21.4925 96.0097 48.0048 96.0097Z"
@@ -20,7 +19,7 @@ const Icon: React.FC<LegacySvgProps> = (props) => {
           <rect width="96" height="96" fill="white" />
         </clipPath>
       </defs>
-    </LegacySvg>
+    </svg>
   );
 };
