+++ src/widgets/WalletModal/config.tsx
@@ -1,20 +1,20 @@
-import Metamask from "./icons/Metamask";
-import Brave from "./icons/Brave";
-import MathWallet from "./icons/MathWallet";
-import TokenPocket from "./icons/TokenPocket";
-import TrustWallet from "./icons/TrustWallet";
-import WalletConnect from "./icons/WalletConnect";
-import BinanceChain from "./icons/BinanceChain";
-import OntoWallet from "./icons/OntoWallet";
-import Nabox from "./icons/Nabox";
-import SafePalWallet from "./icons/SafePalWallet";
-import TorusWallet from "./icons/TorusWallet";
-import Coinbase from "./icons/Coinbase";
-import Unstoppable from "./icons/Unstoppable";
-import Bitkeep from "./icons/Bitkeep";
+import Metamask from "../../components/Svg/WalletsIcons/Metamask";
+import Brave from "../../components/Svg/WalletsIcons/Brave";
+import MathWallet from "../../components/Svg/WalletsIcons/MathWallet";
+import TokenPocket from "../../components/Svg/WalletsIcons/TokenPocket";
+import TrustWallet from "../../components/Svg/WalletsIcons/TrustWallet";
+import WalletConnect from "../../components/Svg/WalletsIcons/WalletConnect";
+import BinanceChain from "../../components/Svg/WalletsIcons/BinanceChain";
+import OntoWallet from "../../components/Svg/WalletsIcons/OntoWallet";
+import Nabox from "../../components/Svg/WalletsIcons/Nabox";
+import SafePalWallet from "../../components/Svg/WalletsIcons/SafePalWallet";
+import TorusWallet from "../../components/Svg/WalletsIcons/TorusWallet";
+import Coinbase from "../../components/Svg/WalletsIcons/Coinbase";
+import Unstoppable from "../../components/Svg/WalletsIcons/Unstoppable";
+import Bitkeep from "../../components/Svg/WalletsIcons/Bitkeep";
+import OKX from "../../components/Svg/WalletsIcons/OKX";
+import MadWallet from "../../components/Svg/WalletsIcons/MadWallet";
 import { Config, ConnectorNames } from "./types";
-import OKX from "./icons/OKX";
-import MadWallet from "./icons/MadWallet";
 
 const connectors: Config[] = [
   {
